requirements.txt
alabaster==0.7.12
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
Cerberus==1.3.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.2
Deprecated==1.2.13
docutils==0.18.1
flake8==4.0.1
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
iniconfig==1.1.1
invoke==1.7.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
pip==22.1.2
pluggy==1.0.0
prettytable==3.3.0
psutil==5.9.1
py==1.11.0
pybind11==2.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-mock==3.8.2
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
setuptools==63.2.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinx-click==4.3.0
tomli==2.0.1
urllib3==1.26.10
wcwidth==0.2.5
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.1
